+++ src/Trans.js
@@ -1,6 +1,6 @@
-import React from 'react';
+import React, { useContext } from 'react';
 import HTML from 'html-parse-stringify2';
-import { getI18n } from './context';
+import { getI18n, getHasUsedI18nextProvider, I18nContext } from './context';
 import { warn, warnOnce } from './utils';
 
 function hasChildren(node) {
@@ -130,7 +130,8 @@ export function Trans({
   t: tFromProps,
   ...additionalProps
 }) {
-  const i18n = i18nFromProps || getI18n();
+  const { i18n: i18nFromContext } = getHasUsedI18nextProvider() ? useContext(I18nContext) : {};
+  const i18n = i18nFromProps || i18nFromContext || getI18n();
   if (!i18n) {
     warnOnce('You will need pass in an i18next instance by using i18nextReactModule');
     return children;