+++ src/index.js
@@ -1,6 +1,8 @@
 export { Trans } from './Trans';
 export { useTranslation } from './useTranslation';
 export { withTranslation } from './withTranslation';
+export { Translation } from './Translation';
+export { I18nextProvider } from './I18nextProvider';
 export { withSSR } from './withSSR';
 export { useSSR } from './useSSR';