+++ src/useSSR.js
@@ -1,7 +1,10 @@
-import { getI18n } from './context';
+import { useContext } from 'react';
+import { getI18n, getHasUsedI18nextProvider, I18nContext } from './context';
 
-export function useSSR(initialI18nStore, initialLanguage) {
-  const i18n = getI18n();
+export function useSSR(initialI18nStore, initialLanguage, props = {}) {
+  const { i18n: i18nFromProps } = props;
+  const { i18n: i18nFromContext } = getHasUsedI18nextProvider() ? useContext(I18nContext) : {};
+  const i18n = i18nFromProps || i18nFromContext || getI18n();
 
   // nextjs / SSR: getting data from next.js or other ssr stack
   if (initialI18nStore && !i18n.initializedStoreOnce) {